+++ tests/Integration/RestAPI/helper.ts
@@ -2,11 +2,12 @@ import { IOptions } from "../../../src/index";
 
 const parameters = require('../../../config/parameters');
 
-export const RUN_INTEGRATION_TESTS = process.env.RUN_INTEGRATION_TESTS === 'true';
+export const RUN_INTEGRATION_TESTS = parameters.runIntegrationTests === 'true';
 
 // api host
 const host = parameters.apiUrl;
 const version = 'v1';
+const organizationUid = parameters.organizationUid;
 // client credentials
 const clientId = parameters.organizationAuth.clientId;
 const secret = parameters.organizationAuth.secret;
@@ -17,6 +18,7 @@ const token = parameters.accountAuth.token;
 export const opts: IOptions = {
 	url: host,
 	version: version,
+	organizationUid,
 	organizationAuth: {
 		clientId,
 		secret,